Change init map to have values of vec1
Summary:
We need these to have all paths instead of just a single path, since when doing lookups, we want to return the `.py`/`.pyi` preference, if there is one that we can fulfill.

Also, when creating the index, we're not sure which file we might be coming from/doing an implicit init search for if there are `.py` and `.pyi` results.

Reviewed By: kinto0

Differential Revision: D85171224

fbshipit-source-id: 2a9eaeb16ea4351a5398167cdf0fad263ef76be1