Add todo to refactor bundles.rs into an implementation using Traits
Summary:
The current refactoring approach is making the actual API a little more confusing since bundled.rs now has to handle the logic for each different stub type. This is ok now but if we add more types in the future this will become difficult to maintain.

We should instead using a Trait so that each different stub type can house its own implementation.

Reviewed By: kinto0

Differential Revision: D85062916

fbshipit-source-id: 6e3592abefe1e412bbc68798f2732d92010dc262