Get base explicit and implicit init files
Summary:
## This diff...

makes a function that will enable us to cache existing `__init__.pyi?` files and synthesize packages for directories without init files. For now, we only synthesize for the directory containing a given file.

## Background

Buck synthesizes `__init__.py` files when they don't exist for all directories containing at least one file with a `.py` or `.pyi` file. So far, Pyrefly has been working under the assumption that users would provide `__init__.py` files or import a module directly when working under a build system.

For the most part, this isn't a problem, but it means any imports pointing to a directory or not relative to the repo root will fail.

Reviewed By: kinto0

Differential Revision: D85171227

fbshipit-source-id: 261831f44ee3ea0bd125585fc54bc38fb6ea891f