Set packages on PythonLibraryManifest
Summary:
## This diff...

combines the functions created in the previous two diffs and fills in packages used by this `PythonLibraryManifest`. We set this when running `TargetManifestDatabase::produce_map()`, after collecting all real init files.

## Background

Buck synthesizes `__init__.py` files when they don't exist for all directories containing at least one file with a `.py` or `.pyi` file. So far, Pyrefly has been working under the assumption that users would provide `__init__.py` files or import a module directly when working under a build system.

For the most part, this isn't a problem, but it means any imports pointing to a directory or not relative to the repo root will fail. We fix this by keeping track of existing packages, and return them

Reviewed By: kinto0

Differential Revision: D85166343

fbshipit-source-id: 78b930c8545f971f13bf4db75da61ebd74444785