Construct ancestor implicit init files
Summary:
## This diff...

takes the result from the previous diff, constructing real and synthesized packages, and continues to synthesize packages up to the directory containing the build file. This enables relative imports for directories referenced by the buildfile

## Background

Buck synthesizes `__init__.py` files when they don't exist for all directories containing at least one file with a `.py` or `.pyi` file. So far, Pyrefly has been working under the assumption that users would provide `__init__.py` files or import a module directly when working under a build system.

For the most part, this isn't a problem, but it means any imports pointing to a directory or not relative to the repo root will fail.

Reviewed By: kinto0

Differential Revision: D85171225

fbshipit-source-id: 2aacd4a8079945154fa82a53ac6dab2c65c6a8fa