Always exclude certain files
Summary:
In a quick discussion, we decided that the default project excludes we're using right now should always be excluded, no matter what. If someone wants to type check stuff in one of these directories, we'll have to make a new flag or something to enable type checking them.

Note, unlike a lot of the other stuff we do, we don't want to absolutize this, since we should probably not include them even if they occur outside of cwd/the config directory.

We also move the functionality of constructing the project excludes field into `ConfigFile::configure()`, so we can use it in `lsp::server::get_diag_if_shown` without having to recompute it every time.

Reviewed By: kinto0

Differential Revision: D85266724

fbshipit-source-id: ff1c5f368e77b0b06017ab372acaccce58a13d21