Add test reproducing autocomplete crash on type with Var
Summary:
We get the base type for completion from Answers::get_type_trace which calls Type::for_display, which converts any unsolved Var to Var::ZERO.

But to get the type of the attributes, we pass the base type into various typing functions (attribute lookup), which do not expect Var::ZERO.

The API between answers and lsp is a bit confused I think. The diff summary in D80365200 suggests that scrubbing out Vars should be ok, because anything "outside of Answers" shouldn't need Var information. But is LSP stuff "outside of Answers"? Not clear.

The next 2 diffs in the stack will fix two issues:
1. Emit the correct error when we see an unexpected Var::Zero. Currently we complain about "var leak" but this is different.
2. Patch over the issue to ignore Var::ZERO in expand_var, which is where this test case is panicking.

Reviewed By: stroxler

Differential Revision: D85280867

fbshipit-source-id: 6e4859dc742b0aff38812a6600da7c3c319c85f4