Fix error message when reaching an unexpected Var::ZERO
Summary:
If we can't look up a Var, then we assume we have a var leak. This makes sense _except_ for Var::ZERO, which will never appear in any variable map.

I found one place where we had a special error, but we can move it into generic code. This fixes a case where we gave the "var leak" error for Var::ZERO.

Reviewed By: stroxler

Differential Revision: D85281202

fbshipit-source-id: 45f0f6911f25c125116f120219fb8983327996ec