Fix crash on unexpected Var::ZERO during autocomplete
Summary:
See D85280867 for details. During autocomplete, we perform some type operations in attribute lookup and end up calling expand_vars, which tries to look up the underlying variable for any Var. But because we are dealing with a "for display" type, we hit the Var::ZERO case.

This diff adds code to replace Var::ZERO with Unknown during expand_vars. This is reasonable, but I it does seem kind of sketchy that we are doing "type stuff" with a type after calling `for_display` on it.

Fixes https://github.com/facebook/pyrefly/issues/1257

Reviewed By: stroxler

Differential Revision: D85281201

fbshipit-source-id: 1199e69cea01249992b6c77765c1ce8550b09043