Add testcase about existing id field
Summary:
By default, the id field is an `int`. But if the field already exists, we skip autogeneration.

In the next diff, we'll add support for that by syhthesizing an id field with a type int (I am looking into whether we will read it from the stubs).

If a field already exists, we skip synthesizing it.

Reviewed By: rchen152

Differential Revision: D85282830

fbshipit-source-id: 2041ad9852dbabd52250eec99097dd91e28b5af5