Separate loop recursive vs general recursive Variable creation
Summary:
I want to represent the loop recursive and general recursive cases differently
so that their behavior can diverge (Why? because loops are quite different from
normal recursion because we have a very good guess about the type, whereas in
the general case we are blind).

As a result I need to separate the APIs for creating them.

Reviewed By: rchen152

Differential Revision: D85178067

fbshipit-source-id: 8a95e6055dbceecd5a9f0369226b4a591a4247c7