Change "default" to "prior_{idx,type}" in solve-stage code
Summary:
The name `default` doesn't mean much here, it's strictly worse than a
name like "prior type" / "prior idx" that actually explains what the
value means.

Moreover, is easily confused with the notion of default type parameters, which
appears in some of the same code (because it's relevant to
`Variable::Quantified`) and where the name "default" is actually an official
part of the spec.

Let's rename it.

Reviewed By: rchen152

Differential Revision: D85178066

fbshipit-source-id: e564f48f28607af4eae1887d63bab6e84b935c20