Rename `base_type_for_newtype`
Summary:
I found `new_type_base_type` a little hard to read - it's not a bad name,
it's accurate, but you can't easily see the "grouping" of parts in the name
to make sense of it.

This name is equivalent but I think is a bit better grouped for readability.

Reviewed By: rchen152

Differential Revision: D85354018

fbshipit-source-id: ba47da724223bd0c9a4bfee82d484d1abb232fb3