Extend testcase to check type for pk
Summary:
Luckily, it looks like we are already handling the types of the pk!
I am not sure if we cover 100% of the cases, but we do handle this basic scenario.

The next step is to extend the implementation to not generate the id field when the custom pk is present.

I haven't determined where this data should live (on the fly vs the metadata) so I will look into this part next.

Reviewed By: rchen152

Differential Revision: D85294539

fbshipit-source-id: eade630c83ccbdede90ba286f757d185385224c2