Only use namespace results in module finding if style filter is none
Summary:
When we're using doing module finding, namespace modules are the lowest priority thing we want to return. When doing module finding in the IDE though, we may miss a `.py` or `.pyi` result because we were applying a `StyleFilter`, and if there's only one of them, return a namespace result incorrectly.

This diff makes it so we only return a namespace result when no style filter is applied.

Reviewed By: kinto0

Differential Revision: D85294549

fbshipit-source-id: 52131b5809516491b4733f06df44f5729a8955d8