Extend pk testcase to check inheritance
Summary:
I think based on the fact that we need to handle inheritance, it's cleaner to store the data we need on the metadata level instead of compute it on the fly.

the data would just be an option<type> of the pk and we would check the data before generating the id field. We also have to assign pk to that custom type.

Reviewed By: rchen152

Differential Revision: D85370739

fbshipit-source-id: b934552c458f05e19db1c4e5622bf058452ca162