always use fallback module path when building lsp modulenames
Summary:
I've been struggling to fix the tests that have to do with workspace-level features not working when there's no config or marker file present.

## Alternatives
My other fix (D85207886) ended up modifying config finder to invent a new type of config with a root at the folder root when workspaces exist. This let us use make a modulename that wasn't `__unknown__` by using the workspace root. But it was complicated, didn't scale well (it only worked for workspace roots, nothing in between).

## This diff

This diff considers all fallback search paths when we are constructing a modulename. This change relies on the fact that any potentially incorrect modulename (by using the first fallback) is better than an unknown modulename. Is that true connernilsen?

Reviewed By: connernilsen

Differential Revision: D85309270

fbshipit-source-id: 6863f687f123896f31773590c8551fa3227cca62