Remove old-to-new error kinds map
Summary:
Last week, I renamed a few error kinds and temporarily made it possible to suppress errors using both the old and new names. Now that we've done a release in which both can be used, drop support for the old names for the next release.

We're still in alpha, so I'm not too worried about making a breaking change =P I did make sure that there weren't many suppressions on GitHub using the old error names.

Reviewed By: stroxler

Differential Revision: D85399946

fbshipit-source-id: f67f5f1e809de601b531cc23a93829e07f8abfa1