Add a way to pass along a FindError when computing module-related information
Summary:
This diff adds some new types and functions that I'll use to make it possible to return both a found import and a FindError from the same lookup.

These are all unused for now.

Context: this stack takes a somewhat long and tortuous route (sorry!) toward adding an `untyped-import` error kind for django. I first do a lot of refactoring to prepare for the new error kind.

Steps:
1. A series of changes to add and use a `ResultWithFindError` type instead of `Result<T, FindError>` when resolving imports, to give us the capability to return both a found import and an error. <= YOU ARE HERE
2. Conversion of `ResultWithFindError` to a more readable `FindingOrError` representation.
3. Use of `FindingOrError` to improve the existing `missing-source` error, as both an opportunistic improvement and a quick verification that we're now able to return both an error and a found import.
4. Addition of `untyped-error` for django.

Reviewed By: migeed-z

Differential Revision: D85407823

fbshipit-source-id: ff5c43124f8253eacfc54dd76e4c769b0db9713b