Make functions that return Result<ModulePath, FindError> return ResultWithFindError<ModulePath> instead
Summary:
As a first step in passing module info alongside a FindError, make functions that return either a ModulePath or a FindError return WithFindError<ModulePath> instead in the error case. For now, this is a no-op, as we always set the `finding` field of WithFindError to None.

Context: this stack takes a somewhat long and tortuous route (sorry!) toward adding an `untyped-import` error kind for django. I first do a lot of refactoring to prepare for the new error kind.

Steps:
1. A series of changes to add and use a `ResultWithFindError` type instead of `Result<T, FindError>` when resolving imports, to give us the capability to return both a found import and an error. <= YOU ARE HERE
2. Conversion of `ResultWithFindError` to a more readable `FindingOrError` representation.
3. Use of `FindingOrError` to improve the existing `missing-source` error, as both an opportunistic improvement and a quick verification that we're now able to return both an error and a found import.
4. Addition of `untyped-error` for django.

Reviewed By: migeed-z

Differential Revision: D85407824

fbshipit-source-id: f02f51a013aac7c5b71a707d41a05002af2c83a2