use workspace root as first fallback heuristic if available
Summary:
### Context
This stack continues my work on workspace-level features when no config is present. This is an often mislooked feature internally (because we expect to always have a config) but is very very important for open source developers who may not have a config.

### Background

D85309270 made it so we use fallback_search_paths as a fallback to building handles in cases where we don't have a config and have no idea where the project bounds start.

This makes find-references work in files at the root of the project:

```
project/
    foo.py
    bar.py
```
when we build the handles for project, we will make `ModuleName`s foo + bar. so when bar imports `foo`, we can find it.

But this doesn't work when we have nested folders:
```
project/
    dir1/
        foo.py
    bar.py # import dir1.foo
```

The handles we create will always use the first fallback search path as a heuristic, which will always make the `ModuleName` the file name!

### This diff
In order for bar to find foo, we should add another heuristic: workspace folder root.

By building modulenames based on this, module names do not overlap and are a little mroe correct than they used to be.

Reviewed By: connernilsen

Differential Revision: D85451961

fbshipit-source-id: 9bc4b3a77fae23c424247149889db6fa4fcac72f