Rename `tuple_base` to `tuple_ancestor`
Summary:
This is non-None if there is any tuple in the inheritance tree,
not just a direct base class.

That means it's an ancestor, not a base. The name is important, otherwise
the semantics won't line up to the vocabulary; this is directly relevant
to a bug we had in the first version of an improvement to NewType behavior;
an external contributor wrote code that would be correct if "tuple_base" actually
meant a base, but it didn't.

Reviewed By: rchen152

Differential Revision: D85470428

fbshipit-source-id: 0042db018fbbf5528b262e1813dd87b51e9764ba