Rename `has_strict` to `has_strict_pydantic_metadata`
Summary:
If someone is trying to maintain the `ClassBases` code in general and
is working on anything other than pydantic, the name `has_strict` won't
mean much to them.

Any name that makes it clear this is a pydantic-specific thing helps, we
immediately have some idea where to look in the code, the tests, and the
diff history to understand it better.

Reviewed By: rchen152

Differential Revision: D85470430

fbshipit-source-id: 3f882d0933fc87a071137db9fc9047cd76f108f4