Add full module and class to function metadata
Summary:
# Problem
When building call graphs for pysa, we want to fully identify a function given a `Type::Function` or `Type::Overload`. While `FuncMetadata` contains the module name and class name, this is not enough to uniquely identify the function, because multiple modules can have the same name, and multiple class can have the same name within a module.

# Solution
Instead of storing `ModuleName` and `Name` for the class, let's store `Module` and `Class`, which fully identify a module and class.
One downside is that it is not possible to create a `FuncId` from a `FunctionKind`, but we can work around that.
This also requires a little bit of boilerplate to define `PartialEq`, `Hash`, `PartialOrd` on `FuncId`.

Reviewed By: stroxler

Differential Revision: D85441657

fbshipit-source-id: 9277543fadf011047024c971f27f912e4bd03238