Track behavior of not-boolable types in boolean operators
Summary:
A group of users is making use of non-callable `__bool__` to prevent
bugs due to sketchy truthiness checks, and asked us to investigate this
case.

Our behavior is conservative - if Python sees `if x or y`, it actually only
has to evaluate `x` as a boolean, not `y`, so we could skip the check.

I put in a TODO for this and will stack a fix; I decided not to mark it as
a bug since I could imagine someone arguing that this behavior is strict
but good; I'm going to fix it since we really don't want false positives
in Pyrefly in cases where the runtime behavior is clear.

Reviewed By: yangdanny97

Differential Revision: D85481910

fbshipit-source-id: e309f739bc8806d8166e272b3042d786027152a9