Organize the (non-operator) `__bool__` test a bit
Summary:
It was excessively commented, with many of the comments just telling
you what was in the next line; as a result it was harder to see the
structure (which is: create some fixtures, test some cases, and
make sure we aren't treating `__getattr__` as implying `__bool__`).

This diff cleans it up a bit; changes are whitespace-only.

Differential Revision: D85451210

fbshipit-source-id: 910bd29c67ea276e2436979de21187294762609a