Add a note to a unit test
Summary:
It's not immediately obvious unless you've looked into it that the runtime
will use the result of `__bool__` and then discard it, using the underlying
instance as the boolean operator result value.

This is sufficiently non-obvious that I got a reach-out with a bug report
complaining that we do the right thing, rather than what the person expected.

I think it's worth explicitly noting that the behavior has been verified against
the runtime in a comment.

Differential Revision: D85451206

fbshipit-source-id: e822d4eb1fcfd951cf566480553cb5b4e8744919