Distribute `__bool__` checks across unions in `check_dunder_bool_is_callable`
Summary:
This diff fixes a gap in Pyrefly: because the semantics of the
`type_of_magic_dunder_attr` function is to drop everything in a union if we
don't find a match in *every* member (which is often the right behavior in
other cases), it was causing us to skip the `__bool__` check entirely when
any member of a union has no `__bool__` defined at all.

This is wrong, since we want to complain if we know there are union member
types for which `__bool__` is non-callable.

This is a bit of a quick fix: the errors are much more correct and the error
messages are much better in union cases now, but ideally we would produce just
a single error instead of duplicates for fat unions. But I'm a bit short on time,
and this behavior is pretty good plus I don't think non-callable `__bool__` is
going to be a very common error so I left a TODO for now.

Reviewed By: yangdanny97

Differential Revision: D85481913

fbshipit-source-id: 27277a929cb9bd786c16ae2093415efde184ae27