Rename some vars
Summary:
The name `condition_type` isn't terrible, but it's a little ambiguous
since the `__bool__` behavior means there are multiple types involved in
how Python views the evaluation of a condition - I think a more verbose
name that explicitly says this is the type of the actual AST term is
clearer; since the function is pretty short, the longer name doesn't hurt
readability.

I think `cond_bool_ty` isn't a great name - this is the type of the `__bool__`
*attribute*, which doesn't really map to the "type of the condition" (which
depending on how you look at it is either `condition_type` or the result of
*calling* `__bool__`).

I'm about to tweak this code to distribute over unions, and this var name
got me a little mixed up. Also, I would like to keep the same name when we
unwrap the `Option` for simplicitly.

Reviewed By: yangdanny97

Differential Revision: D85481915

fbshipit-source-id: 6c594b37346c3af7886ec237cd59026cb6b247ba