Simplify one of the test lines for `__bool__`
Summary:
If we assume that the handling of `True` in ternary expressions
is correct (and this unit test isn't supposed to be verifying that - we
have other tests), this simpler code is equivalent.

This is particularly important because if we *didn't* account for the
truthiness of `True`, this test would be verifying union behavior, but
because of truthiness there's no union here.

That's important because we don't actually support unions, so the existing
test was both overly complex and potentially misleading.

Differential Revision: D85451211

fbshipit-source-id: bf5e29c850ab50626b93b3691cff7d23fedc5d20