Partially clean up `test_valid_dunder_bool`
Summary:
The `__bool__` tests are in a weird state that makes it hard to see
what we support, they are written in a way that is hard to skim because
the blocks aren't clear from blank lines, plus they have a mixture of
incorrect comments and things living in the wrong test.

This diff fixes the first part of `test_valid_dunder_bool` by making
it easier to see the three cases we really want to allow: `Any`, `Never`,
and `__getattr__`; the third of these was improperly put in
`test_invalid_dunder_bool` so I moved it.

The rest of the test needs to be split out, and the comment is completely
wrong (we are in fact checking unions, which the comment says we don't;
I think this has been true since adding `Union` as a direct `AttributeBase`
a few months ago); I'll clean that up in the next diff.

Differential Revision: D85451208

fbshipit-source-id: d0db8bbe86aa1c0584b0d70051817375554ef1a7