Create handler for resolving target from qualified name
Summary:
In patching-related utils,

We are using importlib to resolve a target patch string to its definition.
https://www.internalfb.com/code/fbsource/[4c8e4073e6b99a2b1fd06abdadbc56033b50ca41]/fbcode/instagram-server/distillery/code_migration/context_manager_converter/patch_utils.py?lines=36-45

and we are using resolve_target to figure out the earliest declaration:
https://www.internalfb.com/code/fbsource/[4c8e4073e6b99a2b1fd06abdadbc56033b50ca41]/fbcode/instagram-server/distillery/code_migration/context_manager_converter/patch_utils.py?lines=48%2C64%2C74

This has two problems:
1. The module must be imported for `resolve_target` (if the target is a string, technically you don't even need to import the module). Sometimes, we will end up with an unused import just for resolving the target.
2. ~~Importlib is slow~~ I was just looking at this case and i think it might be dead (?)

To solve the two issues above, the solution suggested by Vlad is to create a new query endpoint that can resolve a string that is the qualified name of a function / method to its definition.

In order to do this, we fabricate a snippet of code from the given qualified name (for example, `a.b.Class.method`):
```
import a.b

X = a.b.Class.method
```

and we use the existing `CalleesWithLocation` mechanism to figure out the target and class name from this snippet of code that uses `a.b.Class.method`.

In this diff, we create the endpoint in igwww.rs and query.rs. In the next diff, I will use this new endpoint in py2hack.

Reviewed By: vladima

Differential Revision: D85271889

fbshipit-source-id: 4ff61da425b840572384a990308542375eb77a46