Improve the `join_types` implementation when the base type directly appears in an all-narrow join
Summary:
If the `base_ty` itself appears in a flow merge, it can lead to
nasty behavior where a `Var` gets pinned to itself or a term involving
itself by our subset check.

This caused a stack overflow in Pyrefly - the new implementation does
an equality comparison before the `is_subset_eq` check in order to rule
this behavior out.

Note that the case we are looking at here only occurs in all-narrowing merges,
i.e. merges where we know no branch reassigned the variable; that's why we
are trying to simplify back to `base_ty` in the first place, and if `base_ty` is
directly part of the join it is always correct to do so.

------

Note: I'm still not 100% sure whether this is a minimal repro of the issue.

I strongly suspect what's going on is that the combination of `break` and conditional
narrows is somehow causing us not to eliminate the recursive binding for `results`,
and then the negative-branch narrows wind up with the recursive `Var` being compared
to itself. But it's actually kind of tricky to print debug, so I'm not entirely confident yet and
even if that is what's happening I don't understand *why* it would prevent loop optimization.

This isn't ideal and I'm hoping to debug some more. I'm relatively confident that we could produce
similar behavior with other examples, but the loop phi optimisation *should* have eliminated
the loop Phi for `results`

Unfortunately I may need a bigger block of time to debug, and I might also need to introduce
some utility functions - we don't have great tools yet for viewing what's inside a `Var`, it would
be good to have a `debugging.rs` helper for that.

In the meantime, we could either land this now or put it off. I'm sure the fix is good, I'm just
not satisfied I understand why we are hitting the problem on this example.

Reviewed By: samwgoldman

Differential Revision: D85594777

fbshipit-source-id: 9d684d8b8b1aff2943d0f7357c87845b1b4598b6