Make `missing-source` respect error configuration
Summary:
Previously, the `missing-source` error ignored all error configuration (e.g., requesting `--error=missing-source` did nothing). Now this error respects configuration. Implementation-wise, this means we always log the error and choose whether to display it or not based on the configured severity level, rather than choosing whether to log it or not based on the value of `--ignore-missing-source`.

For backwards compatibility, I also preserved the behavior of `--ignore-missing-source=false` (enable `missing-source` with severity Error). This does lead to a bit of weirdness when `missing-source` is configured in multiple ways. What I ended up doing is enabling the error if it is explicitly enabled in any way (even if it is also explicitly disabled in a different way!), with the severity defaulting to Error if unspecified.

Maybe we should deprecated `--ignore-missing-source` at some point? (cc connernilsen) But I think the behavior in this diff is good enough for now.

For https://github.com/facebook/pyrefly/issues/1406.

Reviewed By: connernilsen

Differential Revision: D85622655

fbshipit-source-id: 45a67f52306303c84b07ba888a7c185e60b18411