Include third party stubs in update.py script
Summary:
## Stack plan

This stack will be the pieces to actually get Third Party Typeshed stubs into Pyrefly. We have done a lot of work in order to support this, but this diff will do the final stub of actually bundling them with Pyrefly and ensuring that that they work properly with VS code quickfixes and well as CLI warnings.

## Steps of plan
1. Update script(this diff)
2. Bundle third party typeshed with pyrefly.
3. Update finder.rs to properly find the third party typeshed modules.
4. Add quick fixes for VScode and CLI warnings.

Reviewed By: kinto0

Differential Revision: D85054792

fbshipit-source-id: 0a2f8b3af89818ed9fe46f94310585e8038ef9c8