Bundled third party typeshed stubs with Pyrefly.
Summary:
This diff looks massive, but the only actual change here is that we are included the third party stubs in the repo. There should be no functional change here but still good to be wary and make sure nothing has changed in terms of module resolution or any pre-existing behavior.

Before adding the third party stubs the binary size was 1.8MB and after adding the stubs the size has jumped to 8.9MB

Reviewed By: kinto0

Differential Revision: D85442094

fbshipit-source-id: ca8572aabc308a67f1e76015b212909e5c2a61df