refactor DisplayPos to differentiate between source and notebook positions
Summary:
Mapping the positions on-demand is too error-prone, and it's hard to know whether a position refers to a cell level or file level position if the representation is the same. This diff refactors DisplayPos to have 2 variants, for source and notebook positions.

Notebook positions include additional cell-level information (cell number and line number within the cell).

DisplayPos now has 2 APIs to get the line number, one that gets the line within the concatenated source (used for things like extracting snippets, suppression comments, etc), and one that gets the line within the cell (used for things like error messages).

There are probably a few usages that need to change as we build out LSP support, but right now this should be adequate for type checking error messages.

Reviewed By: stroxler

Differential Revision: D85678767

fbshipit-source-id: a36708fc15916437bd5cefcae9e5ec2349b1f454