Add ForeignKey module name
Summary:
We have to recognize a ForeignKey definition, so we should add it to the module name.

The next step is to collect the FK information so that it can be added to the binding layer.

I think a `SmallMap<Name, Box<Expr>>`  where the name is the name of the field and the expr is the type of the foreign key.

Reviewed By: rchen152

Differential Revision: D85697859

fbshipit-source-id: d2c82f08be0c106bd75a3ba5bf1c1de10b6ee28b