Add test for dataclasses.fields() function
Summary:
Adds test_fields_function to verify that calling fields() on both a dataclass type and a dataclass instance does not produce type errors. The test uses assert_type to pin the expected return type as tuple[Field[Any], ...].

I had a report from a user this wasn't working, so worth adding a test so we can be sure it is. We might also want to refine the type eventually.

Reviewed By: yangdanny97

Differential Revision: D85759341

fbshipit-source-id: 472bde1d06c5dbe95736754e9ebd5b9e1960441e