fix hover over error suppressions
Summary:
The original PR didn't have adequate testing, this feature only worked for trailing suppression comments that apply to the current line.

If the suppression appears on a non-code line, it will apply to the next line.

This diff fixes that behavior, makes it work for standalone suppressions, and adds a check to filter by error code.

Reviewed By: stroxler

Differential Revision: D85722892

fbshipit-source-id: 365ea734b15b220f4b12f598e6ceecefe1700b90