Add a comment noting that the Exports indexing trick works only when a pyrefly config file is present
Summary:
I noticed while playing around with the VSCode extension that a pyrefly.toml (even just an empty one) needs to be present for the initial Exports transaction in did_open to do anything useful.

When there's no config file, we have to wait for the slower Indexing transaction to finish. After D85229374 and D85720690, this wait isn't very long (a few seconds in my local testing), so I'm not too worried, but it would be nice to revisit this at some point and see if there's a clever way to speed up the no-config case.

Reviewed By: yangdanny97

Differential Revision: D85726663

fbshipit-source-id: ce64366704fc439f240950891445b7a4648830e1