Properly isolate lsp_interaction tests.
Summary: I noticed something interesting when adding some example Python files for our lsp_interaction tests. Some tests were not properly isolated and were picking up files in other directories which were not intended. We have designated a new directory called basic to act as a fuax-root so that we can have this be the root path for the tests and prevent contamination.

Reviewed By: kinto0

Differential Revision: D85689310

fbshipit-source-id: 6523a2289e517dfac329a4cc0a18ab335d6bc974