Add stdlib path to interpreter query.
Summary:
This diff will set us up to more robustly filter out diagnostic errors for files that are part of the Python standard library. There were several idea on how to do this, with the original being to to filter out files based on their path when compared to a hardcoded list of paths that we knew to be locations of standard library installations. This method proved ineffective since we had to do various checks for different operating systems. It would also be difficult to maintain a fully exhaustive list of all possible paths.

We instead are going to rely on information that we gather from the interpreter. This diff will allow us to do that by doing the following:

1. In fbcode/pyrefly/crates/pyrefly_config/src/environment/environment.rs, we update the Python script in `get_env_from_interpreter` to also get information about what stdlib paths the interpreter is using.
2. We add this to the Python environment struct

In future diffs, we will use this information in order to more effectively filter out errors. We can simply check if the current file is part of any relevant standard library paths by using this new field on the PythonEnvironment rather than having to try and filter out against a non-exhaustive list of potential install locations.

Reviewed By: connernilsen

Differential Revision: D85781296

fbshipit-source-id: 55586c7fb7822cf633ef030a344186e153b0e0df