Check if a pyrefly.toml file is being used when working with typeshed third party stubs
Summary:
This diff does a check in fbcode/pyrefly/pyrefly/lib/commands/config_finder.rs to see if a config file is being used. If they are we should not use third party stubs and ignore any third party stub configuration.

### Rollout plan
The rollout of this feature could potentially introduce a lot of new warnings when the project has a config file. In order to minimize the amount of disruption, we should flag warnings in 2 places:

1. We should have a quick fix in VScode which would will pip install the stubs library or add it to their config file?
2. We should show the same warning in the cli when pyrefly check is run.

We should also add these functionalities before we fully roll out this feature. As discussed in https://fb.workplace.com/groups/657557118131740/permalink/1862227590998014/, we could do a quick add for the IDE, but would not set us up to easily use other types of stubs in the future. If we implement all this now, using other type of stubs in the future will become much easier.

Reviewed By: connernilsen

Differential Revision: D85449733

fbshipit-source-id: c6c2f7c96273514053779fbcd3f11b0fe48d141d