Add new MissingStub FindError and use it in CLI and IDE
Summary:
## Overview

Adds a new FindError that will inform a user when there is a stub available for a package that they do not have installed. The default severity level for this will be a warning.

Also added some tests that cover the error is created properly.

This diff combines with some work Rebecca had done recently and re-uses the `UntypedImport` `ErrorKind` to display these errors. We still introduce a new `FindError` for the missing stubs which we use in `find_import_filtered`, but we map this to the existing `UntypedImport`.

This error will occur when we do have a third party typeshed stub for a particular library AND we have found either a pyrefly.toml file or a pyproject.toml with a pyrefly section. We introduced this error because it is possible that the version of the stubs we have in bundled may be out of date or do not match the version the package that the user has. We would prefer that the user provide their own stubs in this case and prompt them to download their own. More information can be found at https://fb.workplace.com/groups/657557118131740/permalink/1862227590998014/

## Follow ups

As a follow up there are 2 things we can do.

The first is to give the user the actual pip command to run to install the stubs when they are checking through the CLI. Although we do tell them to install the package, we should make the process as frictionless as possible and enable a copy paste.

We could also add a quick fix for the IDE. It may not be possible to actually start the install process for the stubs package, but we could probably do something like copying the command to the users clipboard so they can easily install it themselves.

## Ignoring Django and Typeshed stubs

With this current configuration, both the Django errors and all third party typeshed errors will be ignored due to how noisy they currently are. We need to investigate a way to make these less noisy and make sure that is providing appropriate signal.

Currently here are the number of untyped import errors between Django and other imports for the fbcode/instagram directory

**- Combined: 10, 133**
**- Django: 9,680**
**- Typeshed: 453**

Overall the typeshed errors will only make up about 4% of the total errors if we were to unignore these errors. However, the actual number of errors could still prove to be distruptive.

-----------

## Hover
### With config

{F1983013906}

### Without Config

{F1983013912}

## CLI

### With config

{F1983013919}

### Without config

 {F1983013924}

Reviewed By: stroxler

Differential Revision: D85458062

fbshipit-source-id: 586c1f8599c095299bdedd7908429217ff333a26